requirements.txt
runpod==1.7.7
websocket-server
asyncio